pyproject.toml
[tool.poetry]
name = "emblematic-accounting-agent"
version = "0.1.0"
description = ""
authors = ["Matteo Gevi <matteogevi97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pdfplumber = "^0.11.5"
pytesseract = "^0.3.13"
openai = "^1.59.7"
sentence-transformers = "^3.3.1"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"
forex-python = "^1.8"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.159.0"
pytest = "^8.3.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run-tests = "pytest"
